Remove earlier hack to make cancellation post notifications work
This hack was to make sure requests flushed out of the request queue.
Instead, I am declaring that the correct way to do this is to cancel
the requests using the queue.

Add assert statements to make sure the loadingCount never underflows, which
can cause requests to mysteriously never get sent.